+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Starting from Rust port of the project, all changes will be put into this file.
 
 ## Unreleased (git master)
-
-Nothing yet!
+### Build
+- Change `diacritics` to `secular` because of OSS license issue.
 
 ## [0.4.2] (2024-02-28)
 ### Changes